move `ModelBaseModel` back to `models`
